[BUGFIX] Install Tool: Correct Browser-Title if locked
If the Install Tool is locked, the login-screen states whether the
cause to this is if the ENABLE_INSTALL_TOOL file doesn't exist (and
therefore it's locked) or if no password is set.

Currently, the Browser-Title states "No install tool password set"
even if it's set but the ENABLE_INSTALL_TOOL file doesn't exist.
This Patch sets the correct title in the browser-window:
"The Install Tool is locked".

Resolves: #51408
Release: 6.2
Change-Id: Ia9d25a884667404408a2ef6a375d95f7509d8d9d
Reviewed-on: https://review.typo3.org/23373
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn