[BUGFIX] TCA 'group' selectedListStyle with 'width' breaking layout
The problem came from a row of icons which was always rendered below
the select list with a hardcoded width of "250px", regardless if there
was content or not. Solution is not to render this icon row if there are
no icons, thus allowing the select list to be reduced in width.

Releases: 6.2, 6.1, 6.0
Resolves: #50707
Change-Id: I23520c1c547de1155091e38c69013a18fed9e559