[FEATURE] Enable automatic t3d import for extensions
As preparation for the new distribution (introduction package ...)
handling extensions should be able to provide initial t3d imports.
These are then imported on the root level upon installing the extension.

The t3d file has to be called data.t3d and located in
<extension>/Initialisation/data.t3d.

The data is imported only once (even if you reinstall the extension),
execution state is saved via the core registry.

Resolves: #51437
Releases: 6.2
Change-Id: Ibfe631d173b962781018c6c9e929f4f020e48aa5