[WIP][BUGFIX] File identifiers are case-insensitive
With the default collations used by most databases, casing is ignored.
This leads to problems when searching for an entry in the file database:
If there are two files with the same name in different casings, only one
of the two will be indexed. This index record will be used for both
files, as the database does not see a difference when querying for their
identifiers.

The problem is solved by this commit with the help of a new field
identifier_hash. To also solve the problems arising on case-insensitive
file systems (where all casings are possible and thus the same file
could be indexed multiple times when being created with differently
cased identifiers), file identifiers for these storages are converted to
all-lowercase before hashing.

For existing installations, the new field has to be filled during
upgrade; an upgrade wizard for this is still tbd. Additionally, the
existing indices involving the identifier have to switch to using
identifier_hash.

Releases: 6.2
Change-Id: I805085948e01544efd692498f5e3537612c62050
Resolves: #43027, #46553, 45504