[TASK] Re-enable disabled saltedpasswords task if not finished
If you start a disabled saltedpasswords task it begins to look over all
users again. If it didn't process all users the current position is saved
but there is no message to the backend user that the processing is not
finished. Therefore the task should re-enable itself to be able to
process remaining users.

Resolves: #51465
Releases: 6.2
Change-Id: Ia0abf76af2ba91b68d05f0837dda950660f52c1c
Reviewed-on: https://review.typo3.org/23425
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters