[FEATURE] Enable automatic file import for extensions
As preparation for the new distribution (introduction package ...)
handling extensions should be able to provide initial file imports.
These are then imported to fileadmin/<extensionname>
upon installing the extension.

The files have to be located in <extension>/Initialisation/Files/*.

The files are imported only once (even if you reinstall the extension),
execution state is saved via the core registry.

Resolves: #51466
Releases: 6.2
Change-Id: Icbf63227aa8bbb9fcfb0335bfa5ae92b549a94ba