[FEATURE] Add categories-based menu
Now that system categories are activated by default for pages,
provide a simple way of display categorized pages with a new
typee of menu for the menu/sitemap content element.

The rendering uses a new type of special HMENU called "categories",
which means that it is also possible to create categories-based
menus in pure TypoScript.

Resolves: #51161
Documentation: #51162
Releases: 6.2
Change-Id: I137158f39607ad65dff13a604210a647cb8dda6f