[FEATURE] Capture PHP errors in install tool
Some parts of the install tool can throw fatal PHP errors if
broken extensions are loaded. Currently the database analyzer and
the upgrade wizards are affected by this.

The patch implements a shutdown function that is called even if a
fatal error is thrown and redirects to a special action that renders
the "Check for broken extensions".

Resolves: #51499
Releases: 6.2
Change-Id: Ie64781e59732fb1ef05efd6bd9da1e136241b22e