[TASK] FAL-Drivers: Make getFileInfo more versatile and performant
If a storage needs information about a file, it always has
to retrieve all information about the file from the driver.
This can be very expensive - especially if only the modification
timestamp would be relevant, but sha1 content hash as well
as creation time, access time, size, mimetype ... are detected.
This patch extends the interface with a possibility to only
retrieve a subset of these information. As drivers are only
accessible through an ResourceStorage and the new parameters
are optional this patch does not change behaviour or break
backwards compatiblity.

Releases: 6.2, 6.1, 6.0
Resolves: #51515
Change-Id: I45dec41e3f6cfa2d6c8902b795dc224c729296f8