[FEATURE] Add distribution management to extension manager
To install distribution like the introduction or government
package, the extension manager needs a distribution management.

The distribution management consists of a controller that shows
all available distributions and additions to several actions.
The controller lists the packages to be installed. After
click on installing the distribution extension is downloaded,
all dependencies get resolved via the existing dependency
handling and the user is redirected to the configuration. Upon
saving the configuration the user is redirected to a package
welcome side from where he can jump directly to Web>View and
Web>page.

Resolves: #51537
Releases: 6.2
Change-Id: I4aeecc2d2c2d44bdc138bd502d1b2317d33325fd