[TASK] Categorize sys_file table by default
To be able to migrate from DAM sys_files need to be
categorizable. The core is able to do so, but would
need to be configured manually. To ease migration
this is now activated by default.

If an upgrading installation already modified the
list of categorized tables in its localconf, the
step installer takes care of adding sys_file to
that list silently.

Resolves: #51554
Releases: 6.2
Change-Id: I5a12ddb2ab1da1e7a24b9bfe53af9b0ae33a04a1