[BUGFIX] Fixes possible language handling issue
The language initialization process is currently splitted in 2 main
methods (TypoScriptFrontendController->initLLvars and ->settingLanguage).

Only settingLanguage contains hooks and sets the "sys_language_uid"
property (and "sys_language_content") which determine the displayed
records in Frontend. On the opposite, initLLvars is not hookable and sets
the "lang" property, which may be used by charset conversion methods.

In order to determine display language throught hooks, the only way to do
is to currently call a second time initLLvars after determining language,
but this method populate the "languageDependencies" property without
initializating it, so it becomes filled two times, which lead in incorrect
displayed labels, especially if the displayed language is english and the
default language is not english (as "languageDependencies" is already
filled by "default language")

The 3 parts of the patch are
 * The initLLvars method now pre-initialize the "languageDependencies"
   property
 * The initLLvars is now called from "settingLanguage" method, right
   after the "settingLanguage_preProcess" hook (so it won't be anymore
   needed to call it twice)
 * The convPOSTCharset call is now moved after the language
   initialization, as the language determination done by initLLvars
   can have some impact on charset handling (multi-charset locallang.php
   legacy)

Resolves 49499
Releases 6.2, 6.1, 6.0

Change-Id: I6924345931342d5114b13e5d6fab417387559b9a