[BUGFIX] Recreate processig folder without permission checks
When fetching the processing folder and it does not
exist, it is created on the fly. This should be done
without checking permissions. However by making
storage 0 not browsable, permissions will be checked
when fetching subfolders.

Instead we now fetch the folder objects directly
from the driver, where no permission checks are done.

Additionally change the default permission of
storage 0 to allow reading of folders instead
of reading files. Reading file contents is not
necessary on storage 0, but reading folders is
to be able to check for existing files (e.g.
in processing folder)

Resolves: #51747
Releases: 6.0, 6.1, 6.2
Change-Id: I1a4c3b183323151c436969a478c669adb3856476