[FEATURE] Enable allWrap.splitChar on all includeJS/includeCSS properties
The includeJS/includeCSS properties of the PAGE object feature an
allWrap property which replaces all pipe symbols with the generated
tag. This behaviour differs from how wraps are handled in TYPO3.
A splitChar can now be defined for the allWrap property and the wrap
works like the standard wrap method from the ContentObject class.

Resolves: #51782
Releases: 6.2, 6.1, 6.0
Change-Id: Ic2687394c209b0f58185f37a64b14dff30efd2a1