[BUGFIX] Set absPrefix in GraficalFunctions to PATH_site
To make GifBuilder and GraficalFunctions work in
backend context, it is needed to set the absolute
path prefix for processed images to document root.

It works in FE context, because the current working
directory of PHP is document root, but fails in
BE context, as the CWD can be anything, depending
on the location of the executing script.

Always setting absPrefix to PATH_site will not affect
frontend rendering and will make the FAL file
processing with context crop scape mask also
work in BE context.

Resolves: #51804
Releases: 6.0, 6.1, 6.2
Change-Id: I824333dc6f6cff4286d9eefb20c045e5bb207658