[BUGFIX] TCEmain::log relies on active BE_USER
Several developers abuse t3lib_TCEmain and instantiate it also in other
contexts than BE, for which it never was made.
As this breaks several extensions that use it in FE or CLI mode and the
fix is trivial, avoid crashing.

In case logging is enabled but there exists no BE_USER, simply
prevent logging.

Change-Id: Id5dd43efd21fdca3d18756448d2b760a3f74f286
Resolves: #51829
Releases: 6.2, 6.1, 6.0, 4.7, 4.5