[BUGFIX] Do not add trailing slashes to files
In the security fix #50883 a canonicalize method
for folders has been used where items could
either be files or folders.

This added a trailing slash to files which triggered
a basedir PHP warning if basedir is configured
in PHP.

Use the canonicalize method for file paths instead
to fix this issue.

Resolves: #51761
Releases: 6.0, 6.1, 6.2
Change-Id: I9ef4b6a90ca257ad0104115bd7d48c4de68f135f