[BUGFIX] Enable add markers in LoginController signal
Currently its not possible to have signales handle parameter of slots
as a reference. Because of this its not possible to add marker in the
emit. With this patch the parameter containing the markers is taken
after process and then the markers get handed over to the reference
again.

Change-Id: I00dd75c88e4cabb4a1e5a704e6d4e9a39bef1c7b
Resolves: #51892
Releases: 6.2
Reviewed-on: https://review.typo3.org/24228
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn