[BUGFIX] Configuration controller fatals with local-only extension
If the settings of a local-only extension get saved the configuration
controller fatals. This is because the result of the extension model
repository does not get checked, if it really is a model.

With this fix the instance is first checked and afterwards the category
is read from the model.

Resolves: #51902
Releases: 6.2
Change-Id: I543aa0960f3c2ba1877189b44e47aa7f86c81928