[BUGFIX] Only 1 category field per table possible
It's only possible to do 1 makeCategorizable() call per table,
because, if you do more the later calls will overwrite the
previous category field declarations. That's because of the way
the registry stores the category fields - it can only store 1 field
per table. This change fixes that.
Having more than 1 category field worked in TYPO3 6.0 and
this behaviour occurs since 6.1. So it is considered a bug.

Changes to the tests:
* The tests are adjusted to work with the new registry array
structure.
* The tests doesAddReturnTrueOnDefinedTable and
doesAddReturnFalseOnUndefinedTable were removed
because add() does not return wether the table exists or not
but it returns whether the field was added or not.
* test if registering mor than 1 category field per table works
* test exception if $options is not an array

Resolves: #51921
Releases: 6.2, 6.1
Change-Id: I96d1235176109395339e9841564edadc0c71147d