[FEATURE] Allow multiple category fields per table
It is only possible to do one makeCategorizable() call per
table, because if you do more the later calls would overwrite the
previous category field declaration. That's because of the way
the registry stores the category fields - it can only store one
field per table. This change improves that and allow multiple
fields pers table.

Additionally, a custom label can be set in the options array
for each category field.

Resolves: #51921
Releases: 6.2
Change-Id: I96d1235176109395339e9841564edadc0c71147d