[FEATURE] Allow multiple category fields per table
It was only possible to do one makeCategorizable() call per
table, because if you did more the later calls would overwrite the
previous category field declarations. That's because of the way
the registry stored the category fields - it could only store one
field per table. This change improves that.

Additionally in the options array a custom label can be set for the
categories field.

Resolves: #51921
Releases: 6.2
Change-Id: I96d1235176109395339e9841564edadc0c71147d