[TASK] Remove GFX/im_no_effects setting
Support for Image Magick version 4 and 5 was dropped, allowed
settings for im_version_5 are "im6" and "gm". As soon as one
of this is set for im_version_5, bootstrap automatically sets
"im_no_effects" to 1. In effect, this value is *always* one, so
it is obsolete and can be removed from the system.

Change-Id: Ieb798352aebaae4943018a225518591a952a8bcb
Resolves: #52010
Releases: 6.2
Reviewed-on: https://review.typo3.org/23835
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters