[FEATURE] TypoScript: Allow relative file inclusion
Allows to pass relative paths to INCLUDE_TYPOSCRIPT
if the inclusion is called recursively from a file.

This allows simple nested TypoScript templates that
can be moved or copied without the need to adapt all
includes.

Note: A save in backend with included includes will
save all includes with their absolute path.

Resolves: #52018
Releases: 6.2
Change-Id: Icfb828152076ad285bd145f55f2aabcea89c50d0