[TASK] Remove GFX im_imvMaskState and im_negate_mask settings
There are two very weird setting is GFX array that previously
decided on how "masks" are applied to images if combining
multiple images. GFX/im_negate_mask and GFX/im_imvMaskState are
also manipulated during boostrap and depend on each other.

Fiddling with those settings in Image Magick 6 and Graphics Magick
with its supported versions, it turns out that they are obsolete
if removed completly. The patch does this.

Change-Id: Ic24c877287382f9e718974a0df3577dad2eeaac9
Resolves: #50088
Releases: 6.2