[BUGFIX] Install tool uses empty database settings
Some systems allow to connect with empty username / password
configuration to the database. Althought this is not wrong the install
tool should always ask for database credentials for new installations.
This patch prevents the install tool from setting an empty username and
password as default values without manually submitting the form.

Resolves: #51540, #51956
Releases: 6.2
Change-Id: Ic245635bf649f4ad477d793eda658096ed659aa0