[BUGFIX] Allow BE search in more than one mountpoint
The current implementation of the search is broken for every editor
because just the first mount page is considered. All other subtrees
are completely ignored.

The current implementation allows access to the root page (id=0) if
the search is active. The AbstractDatabaseRecordList uses the function
returnWebmounts() to get the allowed mounts.

Change-Id: Ic462725bb4cba8100c5eeca5f7f47b5711ab2869
Resolves: #35073
Releases: 6.2, 6.1, 6.0