[BUGFIX] Readable Output for type "group" and mm tables is wrong
If a TCA field uses the type "group" and a mm table, a wrong value is
displayed in the list view.
The reason is simple: Support for mm tables is missing. This is only
implemented for the type "group".

To avoid duplicate code the types "select" and "group" are combined in
the switch(

Change-Id: Ia57b7468d493962fc9877f12b685f08eecb012ac
Resolves: #52126
Releases: 6.2, 6.1, 6.0