[FEATURE] Fallback paths (backport from Flow)
With this change it will be possible to define multiple values
for the 'template', 'partial' and 'layout' root path configuration.
Each of the respective options will now have a corresponding setter
that enables configuration of multiple paths to look up when loading a
Fluid template file:
	$view->setTemplateRootPaths(array('first/path', 'second/path', …));
The old setters will be kept and they overrule the fallback paths. So:
	$view->setTemplateRootPath('some/path');
would disable the fallback paths of the previous example. The same is true
for 'setPartialRootPath()' and 'setLayoutRootPath()'.
The rootPath-getters have been deprecated in favor of
'getTemplateRootPaths()', 'getPartialRootPaths()' and
'getLayoutRootPaths()'.

This is a backport of the Flow feature applied in
Change-Id: I530e9a1fadbbd210c980c62cf2022c38fa81bb56 issue #39870

Resolves: #39868
Releases: 6.2
Change-Id: Id5a768ae834c53cd20fd59e762c2acf2ea9e6356
Reviewed-on: https://review.typo3.org/23950
Reviewed-by: Frans Saris
Tested-by: Frans Saris
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring