[FEATURE] Add brute force protection to TYPO3 backend
TYPO3 misses a brute force protection for the backend login. Login
attempts for remote attackers should be temporarily disabled, if a
configurable amount of login failures is reached.

Note: This patch is not completely finished and should be seen as
a discussion base (Core members/security team) for this new feature
in TYPO3.

Resolves: #52170
Releases: 6.2
Change-Id: Id1bd7e35a649ee690add7c4445c63eb7be80b087