[BUGFIX] Revert BackendUtility::BEenableFields in FAL
In #51562 the method BackendUtility::BEenableFields was introduced in FAL.

This has a drawback for third party extensions which are introducing
additional enabledFields for a File. For instance, Media has the "hidden"
flag enabled. Consequently, FAL is not capable of finding a
File that has hidden = 1 anymore; $fileRepository->findByUid(123) is
raising exception #1314354065 whenever file "123" is marked as hidden.

Since we don't have Query Settings like in Extbase, the patch
removes this constraint following the List module practice which
displays all records except the deleted ones.

Change-Id: I91105b2c8484a2382e86bda0b3a9f3e29c7a7495
Resolves: #52245
Resolves: #52236
Releases: 6.0, 6.1, 6.2