[BUGFIX] Alternative implementations for view helpers does not work
Extbase allows to register alternative implementation for
objects. However that does not work for view helpers using
a closing tag. The resolved (alternative) object is compared
to the name of the original view helper and throws an
exception like:

  #1224485398: Templating tags not properly nested. Expected:
  "AlternativeViewHelper"; Actual: "OriginalViewHelper"

A simple, but maybe expensive solution on many alternative
view helpers, is to create an instance of the original view
helper class that will be resolved to the alternative
implementation by the ObjectContainer.

Change-Id: Ie49e5e83c779b4748dc2059f8fbc85552ce4b406
Fixes: #52272
Releases: 6.0, 6.1, 6.2