[WIP][TASK] Unify data for jump URL hash generation
The jump URL generation is unified and moved to a new JumpUrlUtility
class. All methods in the ContentObjectRenderer are now using the same
method for creating jump URLs.

The handling of file and directory names with special characters is
improved and FAL file records with special characters in their file names
will also be URL encoded.

Additionally Unit Tests are added to verify the jump URL function.

WIP: Remove the jumpUrlPatchApplied() check in the
ContentObjectRendererTest. This was added to compare the old and the new
version of the jump URL feature.

Resolves: #52156
Releases: 6.2
Change-Id: I7d2d8cec005aae56a118850071ca669ef9740c9d