[BUGFIX] ObjectManager mut not resolve parameters set explicitely NULL
ObjectManager tries to resolve all optional constructor parameters, no
no matter what the call value dictates. In case the type of parameter
is an interface, and no value is passed, the class is resolved as
parameter type without the term 'interface'. This is how the magical
injection of the *Manager* classes can happen.
But there are cases where no class exists that matches the interface
name. In this case, if NULL  is passed from the calling
method, reflection fails and an exception is thrown.
This exception needs to be caught and the value (NULL) passed to the
contstructor anyway.

Change-Id: I646e6928cbadb486b1e2734efbe1ea5fbd953765
Resolves: #52349
Releases: 6.2