[FEATURE] Add possibility to use an extension as repository
Sometimes extensions depend on customized versions of other
extensions or on extensions that aren't released to the TER.

This feature allows an extension to deliver it's own dependencies
by having the unpacked extensions inside the folder
Initialisation/Extensions. The extensions are then copied to
typo3conf/ext when resolving the dependencies.

Resolves: #51835
Releases: 6.2
Change-Id: I66f6e5d5d40e11471f9766ac2ed9d0eb5386eb49