[FEATURE] ext:install Relaxed permission check
The document root folder may not have correct permissions and
they can not be fixed if the owner is not the executing webserver
user.

This is a common issue during installation, currently the first
step breaks if permission on document root are not 2770 and the
folder is not owned by the executing webserver user.

The patch adds a new option "targetPermissionRelaxed" to
DirectoryNodes and uses it on the RootNode to lower the severity
if permissions are not correct. The installation will proceed
then as long as sub folders can be created.

Change-Id: Ibde9f7ff0dde14502ac24fdb73c0a2ed384119a6
Resolves: #52356
Releases: 6.2