[FEATURE] ext:install Relaxed permission check
The document root folder does not neccessarily have correct
permissions. If for example the folder is not *owned* by the
webserver user, permissions can not be fixed, but nodes within
the document root could be very well created.

This is a common issue during installation, currently the first
step breaks if permission on document root are not 2770 and the
folder is not owned by the executing webserver user.

The patch adds a new option "targetPermissionRelaxed" to
DirectoryNodes and uses on the RootNode to lower the severity
if permissions are not correct and can not be fixed.

Change-Id: Ibde9f7ff0dde14502ac24fdb73c0a2ed384119a6
Resolves: #52349
Releases: 6.2