[FEATURE] Control visibility of modules in the BE
Currently there is no way for controlling the visibility
of a module in the BE. The patch introduces User TSconfig
which can be used to hide modules from the main menu.

options.hideModules = file, help

options.hideModules.web := addToList(func,info)

options.hideModules.system = BelogLog

The change only influences the visibility of modules in the menu
and does not control the access (permission) which still must be
done in the ACL of a BE User / Group as usually.

Change-Id: I6c2ec2c389a154440932afa16f9ca3a2cb4b8116
Releases: 6.2
Resolves: #52482
Documentation: #52692
Reviewed-on: https://review.typo3.org/24274
Reviewed-by: Steffen Müller
Tested-by: Steffen Müller
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring